Define outcome for same sex marriages in Italy.
This commit defines an outcome for same sex marriages in Italy.

Changes also made to the marriage abroad flow to properly make use of this outcome.

Update made to the test response and expected results YAML file.

On running the regression tests, I found no changes to the existing regression test artefacts.